修复bug 金币系数改为double

+++ Assets/Scripts/Local/Manange/PlayInfor.cs
@@ -105,7 +105,7 @@ public class PlayerInfo : IComparable<PlayerInfo>
     }
 
     // Method to add coins
-    public void AddCoins(int amount)
+    public void AddCoins(long amount)
     {
         coinNum += amount;
         if(GameFlowManager.Instance.currentLevelIndex == 0)
